adding things

+++ posts/views.py
@@ -6,7 +6,7 @@ from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated
 from rest_framework import status
 from .models import Post
-from .serializers import PostSerializer, CommunitySerializer, UserSerializer
+from .serializers import PostSerializer
 
 @api_view(['GET'])
 @permission_classes([IsAuthenticated])
@@ -38,6 +38,7 @@ def create_post(request):
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['PUT'])
+@permission_classes([IsAuthenticated])
 def edit_post(request, pk):
     post = get_object_or_404(Post, pk=pk)
     serializer = PostSerializer(post, data=request.data)
@@ -47,7 +48,10 @@ def edit_post(request, pk):
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
 def delete_post(request, pk):
     post = get_object_or_404(Post, pk=pk)
+    if post.user_id != request.user.id:
+        return Response({'message': 'Anda tidak memiliki izin untuk menghapus postingan ini'}, status=status.HTTP_403_FORBIDDEN)
     post.delete()
     return Response({'message': 'Post berhasil dihapus'}, status=status.HTTP_204_NO_CONTENT)
